Move cdb2api hndl into own file
